Moved get-pokemon to use route
