+++ src/rest-api.ts
@@ -1,34 +1,12 @@
-import axios from "axios";
 import bodyParser = require("body-parser");
 import * as express from "express";
-import { NextFunction, Request, Response } from "express";
 import { join } from "path";
 import * as swaggerUi from "swagger-ui-express";
 import * as YAML from "yamljs";
-import { getPokemonErrorMiddleware } from "./middleware/get-pokemon-error-middleware";
-import { PokemonService } from "./pokemon-service/pokemon-service";
+import { getPokemonRoute } from "./routes/get-pokemon";
 import { internalHealthCheck } from "./routes/internal-health-check";
 import { internalSwagger } from "./routes/internal-swagger";
 const swaggerDocument = YAML.load(join(__dirname, "swagger.yaml"));
-
-export const getPokemon = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<void> => {
-  const pokemonNameOrId = req.params.pokemonNameOrId;
-  const pokemonService = new PokemonService(
-    axios.create({ baseURL: "https://pokeapi.co/api/v2" })
-  );
-
-  try {
-    const pokemonInfo = await pokemonService.getPokemonInfo(pokemonNameOrId);
-    res.status(200).json({ pokemonInfo });
-    return next();
-  } catch (error) {
-    return next(error);
-  }
-};
 
 export const createRestApp = () => {
   const app = express();
@@ -40,9 +18,7 @@ export const createRestApp = () => {
   );
   app.use("/internal/health-check", internalHealthCheck());
   app.use("/internal/swagger.yaml", internalSwagger());
-
-  app.use(getPokemonErrorMiddleware);
-  app.get("/api/pokemon/:pokemonNameOrId", getPokemon);
+  app.use("/api/pokemon/", getPokemonRoute());
 
   const port = process.env.PORT || 3000;
   // tslint:disable-next-line: no-console
